[REFACTOR] Renamed samples (#55)
* renamed samples

* renamed scene name

* Removed unused files
