+++ Assets/Scenes/samplescene.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 13b933ed50aa0ff4e8bf942522390a9e
+guid: 0aae421965c7db245a4737db694b379b
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
