+++ Assets/Scenes/samplescene/LightingData.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 56e85d5dc4d99b24f9b01a6f086bd74d
+guid: 59d415d885d26374eaaf986accb2c3ba
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 112000000
