--- Assets/Scripts/UIController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 27247405836299a49abf76b3eb56d895
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
