+++ Packages/com.unity.webrtc/Samples/Example/DataChannelSample.cs
@@ -2,20 +2,29 @@
 using UnityEngine;
 using UnityEngine.UI;
 using Unity.WebRTC;
+using System;
 
-public class Main : MonoBehaviour
+public class DataChannelSample : MonoBehaviour
 {
-#pragma warning disable 0649
+    #pragma warning disable 0649
     [SerializeField] private Button callButton;
-    [SerializeField] private Button hangupButton;
-#pragma warning restore 0649
+    [SerializeField] private Button sendButton;
+    [SerializeField] private InputField textSend;
+    [SerializeField] private InputField textReceive;
+    #pragma warning restore 0649
 
     private RTCPeerConnection pc1, pc2;
+    private RTCDataChannel dataChannel, remoteDataChannel;
     private Coroutine sdpCheck;
+    private string msg;
     private DelegateOnIceConnectionChange pc1OnIceConnectionChange;
     private DelegateOnIceConnectionChange pc2OnIceConnectionChange;
     private DelegateOnIceCandidate pc1OnIceCandidate;
     private DelegateOnIceCandidate pc2OnIceCandidate;
+    private DelegateOnMessage onDataChannelMessage;
+    private DelegateOnOpen onDataChannelOpen;
+    private DelegateOnClose onDataChannelClose;
+    private DelegateOnDataChannel onDataChannel;
 
     private RTCOfferOptions OfferOptions = new RTCOfferOptions
     {
@@ -32,9 +41,7 @@ public class Main : MonoBehaviour
     private void Awake()
     {
         WebRTC.Initialize();
-
         callButton.onClick.AddListener(() => { StartCoroutine(Call()); });
-        hangupButton.onClick.AddListener(() => { Hangup(); });
     }
 
     private void OnDestroy()
@@ -45,24 +52,19 @@ public class Main : MonoBehaviour
     private void Start()
     {
         callButton.interactable = true;
+
         pc1OnIceConnectionChange = new DelegateOnIceConnectionChange(state => { OnIceConnectionChange(pc1, state); });
         pc2OnIceConnectionChange = new DelegateOnIceConnectionChange(state => { OnIceConnectionChange(pc2, state); });
-        pc1OnIceCandidate = new DelegateOnIceCandidate(candidate => { OnIceCandidate(pc2, candidate); });
+        pc1OnIceCandidate = new DelegateOnIceCandidate(candidate => { OnIceCandidate(pc1, candidate); });
         pc2OnIceCandidate = new DelegateOnIceCandidate(candidate => { OnIceCandidate(pc1, candidate); });
-    }
-
-    private IEnumerator Loop()
-    {
-        while (true)
+        onDataChannel = new DelegateOnDataChannel(channel =>
         {
-            yield return new WaitForSeconds(1);
-
-            var desc = pc1.GetLocalDescription();
-            Debug.Log($"pc1 sdp: {desc.sdp}");
-
-            //var conf = pc1.GetConfiguration();
-            //Debug.Log($"ice servers count {conf.iceServers.Length}");
-        }
+            remoteDataChannel = channel;
+            remoteDataChannel.OnMessage = onDataChannelMessage;
+        });
+        onDataChannelMessage = new DelegateOnMessage(bytes => { textReceive.text = System.Text.Encoding.UTF8.GetString(bytes); });
+        onDataChannelOpen = new DelegateOnOpen(()=> { sendButton.interactable = true; });
+        onDataChannelClose = new DelegateOnClose(() => { sendButton.interactable = false; });
     }
 
     RTCConfiguration GetSelectedSdpSemantics()
@@ -74,36 +76,6 @@ public class Main : MonoBehaviour
         };
 
         return config;
-    }
-    IEnumerator Call()
-    {
-        callButton.interactable = false;
-        hangupButton.interactable = true;
-
-        Debug.Log("GetSelectedSdpSemantics");
-        var configuration = GetSelectedSdpSemantics();
-        pc1 = new RTCPeerConnection(ref configuration);
-        Debug.Log("Created local peer connection object pc1");
-        pc1.OnIceCandidate = pc1OnIceCandidate;
-        pc1.OnIceConnectionChange = pc1OnIceConnectionChange;
-        pc2 = new RTCPeerConnection(ref configuration);
-        Debug.Log("Created remote peer connection object pc2");
-        pc2.OnIceCandidate = pc2OnIceCandidate;
-        pc2.OnIceConnectionChange = pc2OnIceConnectionChange;
-
-        Debug.Log("pc1 createOffer start");
-        var op = pc1.CreateOffer(ref OfferOptions);
-        yield return op;
-
-        if (!op.isError)
-        {
-            yield return StartCoroutine(OnCreateOfferSuccess(op.desc));
-        }
-        else
-        {
-            OnCreateSessionDescriptionError(op.error);
-        }
-        sdpCheck = StartCoroutine(Loop());
     }
     void OnIceConnectionChange(RTCPeerConnection pc, RTCIceConnectionState state)
     {
@@ -137,10 +109,71 @@ public class Main : MonoBehaviour
                 break;
         }
     }
+    void Pc1OnIceConnectinChange(RTCIceConnectionState state)
+    {
+        OnIceConnectionChange(pc1, state);
+    }
+    void Pc2OnIceConnectionChange(RTCIceConnectionState state)
+    {
+        OnIceConnectionChange(pc2, state);
+    }
+
+    void Pc1OnIceCandidate(RTCIceCandidate candidate)
+    {
+        OnIceCandidate(pc1, candidate);
+    }
+    void Pc2OnIceCandidate(RTCIceCandidate candidate)
+    {
+        OnIceCandidate(pc2, candidate);
+    }
+
+    IEnumerator Call()
+    {
+        callButton.interactable = false;
+        Debug.Log("GetSelectedSdpSemantics");
+        var configuration = GetSelectedSdpSemantics();
+        pc1 = new RTCPeerConnection(ref configuration);
+        Debug.Log("Created local peer connection object pc1");
+        pc1.OnIceCandidate = pc1OnIceCandidate;
+        pc1.OnIceConnectionChange = pc1OnIceConnectionChange;
+        pc2 = new RTCPeerConnection(ref configuration);
+        Debug.Log("Created remote peer connection object pc2");
+        pc2.OnIceCandidate = pc2OnIceCandidate;
+        pc2.OnIceConnectionChange = pc2OnIceConnectionChange;
+        pc2.OnDataChannel = onDataChannel;
+
+        RTCDataChannelInit conf = new RTCDataChannelInit(true);
+        dataChannel = pc1.CreateDataChannel("data", ref conf);
+        dataChannel.OnOpen = onDataChannelOpen;
+
+        Debug.Log("pc1 createOffer start");
+        var op = pc1.CreateOffer(ref OfferOptions);
+        yield return op;
+
+        if (!op.isError)
+        {
+            yield return StartCoroutine(OnCreateOfferSuccess(op.desc));
+        }
+        else
+        {
+            OnCreateSessionDescriptionError(op.error);
+        }
+    }
+
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <param name="pc"></param>
+    /// <param name="streamEvent"></param>
     void OnIceCandidate(RTCPeerConnection pc, RTCIceCandidate​ candidate)
     {
         GetOtherPc(pc).AddIceCandidate(ref candidate);
         Debug.Log($"{GetName(pc)} ICE candidate:\n {candidate.candidate}");
+    }
+
+    public void SendMsg()
+    {
+        dataChannel.Send(textSend.text);
     }
     string GetName(RTCPeerConnection pc)
     {
@@ -238,19 +271,18 @@ public class Main : MonoBehaviour
         }
     }
 
+    void OnAddIceCandidateSuccess(RTCPeerConnection pc)
+    {
+        Debug.Log($"{GetName(pc)} addIceCandidate success");
+    }
+
+    void OnAddIceCandidateError(RTCPeerConnection pc, RTCError error)
+    {
+        Debug.Log($"{GetName(pc)} failed to add ICE Candidate: ${error}");
+    }
+
     void OnCreateSessionDescriptionError(RTCError e)
     {
 
-    }
-    void Hangup()
-    {
-        Debug.Log("Ending call");
-        StopCoroutine(sdpCheck);
-        pc1.Close();
-        pc2.Close();
-        pc1 = null;
-        pc2 = null;
-        hangupButton.interactable = false;
-        callButton.interactable = true;
     }
 }
