+++ Packages/com.unity.webrtc/Samples/Example/MediaStreamSample.cs
@@ -1,19 +1,25 @@
 ﻿using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UI;
 using Unity.WebRTC;
-using System;
+using UnityEngine.UI;
+using System.Text;
+using UnityEngine.SceneManagement;
 
-public class TransmitText : MonoBehaviour
+public class MediaStreamSample : MonoBehaviour
 {
-    #pragma warning disable 0649
+#pragma warning disable 0649
     [SerializeField] private Button callButton;
-    [SerializeField] private Button sendButton;
-    [SerializeField] private InputField textSend;
-    [SerializeField] private InputField textReceive;
-    #pragma warning restore 0649
+    [SerializeField] private Button addTracksButton;
+    [SerializeField] private Button removeTracksButton;
+    [SerializeField] private Camera cam;
+    [SerializeField] private InputField infoText;
+    [SerializeField] private RawImage RtImage;
+#pragma warning restore 0649
 
     private RTCPeerConnection pc1, pc2;
+    private List<RTCRtpSender> pc1Senders, pc2Senders;
+    private Unity.WebRTC.MediaStream audioStream, videoStream;
     private RTCDataChannel dataChannel, remoteDataChannel;
     private Coroutine sdpCheck;
     private string msg;
@@ -21,10 +27,10 @@ public class TransmitText : MonoBehaviour
     private DelegateOnIceConnectionChange pc2OnIceConnectionChange;
     private DelegateOnIceCandidate pc1OnIceCandidate;
     private DelegateOnIceCandidate pc2OnIceCandidate;
-    private DelegateOnMessage onDataChannelMessage;
-    private DelegateOnOpen onDataChannelOpen;
-    private DelegateOnClose onDataChannelClose;
-    private DelegateOnDataChannel onDataChannel;
+    private DelegateOnTrack pc2Ontrack;
+    private DelegateOnNegotiationNeeded pc1OnNegotiationNeeded;
+    private StringBuilder trackInfos;
+    private bool videoUpdateStarted = false;
 
     private RTCOfferOptions OfferOptions = new RTCOfferOptions
     {
@@ -41,30 +47,40 @@ public class TransmitText : MonoBehaviour
     private void Awake()
     {
         WebRTC.Initialize();
-        callButton.onClick.AddListener(() => { StartCoroutine(Call()); });
+        callButton.onClick.AddListener(() => { Call(); });
+        addTracksButton.onClick.AddListener(() => { AddTracks(); });
+        removeTracksButton.onClick.AddListener(() => { RemoveTracks(); });
     }
 
     private void OnDestroy()
     {
+        Audio.Stop();
         WebRTC.Finalize();
     }
 
     private void Start()
     {
+        trackInfos = new StringBuilder();
+        pc1Senders = new List<RTCRtpSender>();
+        pc2Senders = new List<RTCRtpSender>();
         callButton.interactable = true;
 
         pc1OnIceConnectionChange = new DelegateOnIceConnectionChange(state => { OnIceConnectionChange(pc1, state); });
         pc2OnIceConnectionChange = new DelegateOnIceConnectionChange(state => { OnIceConnectionChange(pc2, state); });
         pc1OnIceCandidate = new DelegateOnIceCandidate(candidate => { OnIceCandidate(pc1, candidate); });
         pc2OnIceCandidate = new DelegateOnIceCandidate(candidate => { OnIceCandidate(pc1, candidate); });
-        onDataChannel = new DelegateOnDataChannel(channel =>
+        pc2Ontrack = new DelegateOnTrack(e => { OnTrack(pc2, e); });
+        pc1OnNegotiationNeeded = new DelegateOnNegotiationNeeded(() => { StartCoroutine(Pc1OnNegotiationNeeded()); });
+        if (!WebRTC.HWEncoderSupport)
         {
-            remoteDataChannel = channel;
-            remoteDataChannel.OnMessage = onDataChannelMessage;
-        });
-        onDataChannelMessage = new DelegateOnMessage(bytes => { textReceive.text = System.Text.Encoding.UTF8.GetString(bytes); });
-        onDataChannelOpen = new DelegateOnOpen(()=> { sendButton.interactable = true; });
-        onDataChannelClose = new DelegateOnClose(() => { sendButton.interactable = false; });
+            addTracksButton.interactable = false;
+            callButton.interactable = false;
+            infoText.text = "Current GPU doesn't support Nvidia Encoder";
+        }
+        else
+        {
+            infoText.text = "Current GPU supports Nvidia Encoder";
+        }
     }
 
     RTCConfiguration GetSelectedSdpSemantics()
@@ -109,6 +125,21 @@ public class TransmitText : MonoBehaviour
                 break;
         }
     }
+    IEnumerator Pc1OnNegotiationNeeded()
+    {
+        Debug.Log("pc1 createOffer start");
+        var op = pc1.CreateOffer(ref OfferOptions);
+        yield return op;
+
+        if (!op.isError)
+        {
+            yield return StartCoroutine(OnCreateOfferSuccess(op.desc));
+        }
+        else
+        {
+            OnCreateSessionDescriptionError(op.error);
+        }
+    }
     void Pc1OnIceConnectinChange(RTCIceConnectionState state)
     {
         OnIceConnectionChange(pc1, state);
@@ -126,8 +157,44 @@ public class TransmitText : MonoBehaviour
     {
         OnIceCandidate(pc2, candidate);
     }
+    public void AddTracks() 
+    {
+        foreach (var track in audioStream.GetTracks())
+        {
+            pc1Senders.Add (pc1.AddTrack(track));  
+        }
+        foreach(var track in videoStream.GetTracks())
+        {
+            pc1Senders.Add(pc1.AddTrack(track));
+        }
+        if(!videoUpdateStarted)
+        {
+            StartCoroutine(WebRTC.Update());
+            videoUpdateStarted = true;
+        }
+        addTracksButton.interactable = false;
+        removeTracksButton.interactable = true;
+    }
 
-    IEnumerator Call()
+    public void RemoveTracks()
+    {
+        foreach(var sender in pc1Senders)
+        {
+            pc1.RemoveTrack(sender);
+        }
+        foreach (var sender in pc2Senders)
+        {
+            pc2.RemoveTrack(sender);
+        }
+        pc1Senders.Clear();
+        pc2Senders.Clear();
+        addTracksButton.interactable = true;
+        removeTracksButton.interactable = false;
+        trackInfos.Clear();
+        infoText.text = "";
+    }
+
+    void Call()
     {
         callButton.interactable = false;
         Debug.Log("GetSelectedSdpSemantics");
@@ -136,28 +203,19 @@ public class TransmitText : MonoBehaviour
         Debug.Log("Created local peer connection object pc1");
         pc1.OnIceCandidate = pc1OnIceCandidate;
         pc1.OnIceConnectionChange = pc1OnIceConnectionChange;
+        pc1.OnNegotiationNeeded = pc1OnNegotiationNeeded;
         pc2 = new RTCPeerConnection(ref configuration);
         Debug.Log("Created remote peer connection object pc2");
         pc2.OnIceCandidate = pc2OnIceCandidate;
         pc2.OnIceConnectionChange = pc2OnIceConnectionChange;
-        pc2.OnDataChannel = onDataChannel;
+        pc2.OnTrack = pc2Ontrack;
 
         RTCDataChannelInit conf = new RTCDataChannelInit(true);
         dataChannel = pc1.CreateDataChannel("data", ref conf);
-        dataChannel.OnOpen = onDataChannelOpen;
-
-        Debug.Log("pc1 createOffer start");
-        var op = pc1.CreateOffer(ref OfferOptions);
-        yield return op;
-
-        if (!op.isError)
-        {
-            yield return StartCoroutine(OnCreateOfferSuccess(op.desc));
-        }
-        else
-        {
-            OnCreateSessionDescriptionError(op.error);
-        }
+        audioStream = Audio.CaptureStream();
+        videoStream = cam.CaptureStream(1280, 720);
+        RtImage.texture = cam.targetTexture;
+ 
     }
 
     /// <summary>
@@ -170,11 +228,15 @@ public class TransmitText : MonoBehaviour
         GetOtherPc(pc).AddIceCandidate(ref candidate);
         Debug.Log($"{GetName(pc)} ICE candidate:\n {candidate.candidate}");
     }
-
-    public void SendMsg()
+    void OnTrack(RTCPeerConnection pc, RTCTrackEvent e)
     {
-        dataChannel.Send(textSend.text);
+        pc2Senders.Add(pc.AddTrack(e.Track));
+        trackInfos.Append($"{GetName(pc)} receives remote track:\r\n");
+        trackInfos.Append($"Track kind: {e.Track.Kind}\r\n");
+        trackInfos.Append($"Track id: {e.Track.Id}\r\n");
+        infoText.text = trackInfos.ToString();
     }
+
     string GetName(RTCPeerConnection pc)
     {
         return (pc == pc1) ? "pc1" : "pc2";
@@ -225,7 +287,7 @@ public class TransmitText : MonoBehaviour
         }
         else
         {
-            OnCreateSessionDescriptionError(op3.error);
+            OnCreateSessionDescriptionError(op3.error); 
         }
     }
