+++ Packages/com.unity.webrtc/Samples/Example/PeerConnectionSample.cs
@@ -1,36 +1,21 @@
 ﻿using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using Unity.WebRTC;
 using UnityEngine.UI;
-using System.Text;
-using UnityEngine.SceneManagement;
+using Unity.WebRTC;
 
-public class AddMediaStream : MonoBehaviour
+public class PeerConnectionSample : MonoBehaviour
 {
 #pragma warning disable 0649
     [SerializeField] private Button callButton;
-    [SerializeField] private Button addTracksButton;
-    [SerializeField] private Button removeTracksButton;
-    [SerializeField] private Camera cam;
-    [SerializeField] private InputField infoText;
-    [SerializeField] private RawImage RtImage;
+    [SerializeField] private Button hangupButton;
 #pragma warning restore 0649
 
     private RTCPeerConnection pc1, pc2;
-    private List<RTCRtpSender> pc1Senders, pc2Senders;
-    private MediaStream audioStream, videoStream;
-    private RTCDataChannel dataChannel, remoteDataChannel;
     private Coroutine sdpCheck;
-    private string msg;
     private DelegateOnIceConnectionChange pc1OnIceConnectionChange;
     private DelegateOnIceConnectionChange pc2OnIceConnectionChange;
     private DelegateOnIceCandidate pc1OnIceCandidate;
     private DelegateOnIceCandidate pc2OnIceCandidate;
-    private DelegateOnTrack pc2Ontrack;
-    private DelegateOnNegotiationNeeded pc1OnNegotiationNeeded;
-    private StringBuilder trackInfos;
-    private bool videoUpdateStarted = false;
 
     private RTCOfferOptions OfferOptions = new RTCOfferOptions
     {
@@ -47,39 +32,36 @@ public class AddMediaStream : MonoBehaviour
     private void Awake()
     {
         WebRTC.Initialize();
-        callButton.onClick.AddListener(() => { Call(); });
-        addTracksButton.onClick.AddListener(() => { AddTracks(); });
-        removeTracksButton.onClick.AddListener(() => { RemoveTracks(); });
+
+        callButton.onClick.AddListener(() => { StartCoroutine(Call()); });
+        hangupButton.onClick.AddListener(() => { Hangup(); });
     }
 
     private void OnDestroy()
     {
-        Audio.Stop();
         WebRTC.Finalize();
     }
 
     private void Start()
     {
-        trackInfos = new StringBuilder();
-        pc1Senders = new List<RTCRtpSender>();
-        pc2Senders = new List<RTCRtpSender>();
         callButton.interactable = true;
-
         pc1OnIceConnectionChange = new DelegateOnIceConnectionChange(state => { OnIceConnectionChange(pc1, state); });
         pc2OnIceConnectionChange = new DelegateOnIceConnectionChange(state => { OnIceConnectionChange(pc2, state); });
-        pc1OnIceCandidate = new DelegateOnIceCandidate(candidate => { OnIceCandidate(pc1, candidate); });
+        pc1OnIceCandidate = new DelegateOnIceCandidate(candidate => { OnIceCandidate(pc2, candidate); });
         pc2OnIceCandidate = new DelegateOnIceCandidate(candidate => { OnIceCandidate(pc1, candidate); });
-        pc2Ontrack = new DelegateOnTrack(e => { OnTrack(pc2, e); });
-        pc1OnNegotiationNeeded = new DelegateOnNegotiationNeeded(() => { StartCoroutine(Pc1OnNegotiationNeeded()); });
-        if (!WebRTC.HWEncoderSupport)
+    }
+
+    private IEnumerator Loop()
+    {
+        while (true)
         {
-            addTracksButton.interactable = false;
-            callButton.interactable = false;
-            infoText.text = "Current GPU doesn't support Nvidia Encoder";
-        }
-        else
-        {
-            infoText.text = "Current GPU supports Nvidia Encoder";
+            yield return new WaitForSeconds(1);
+
+            var desc = pc1.GetLocalDescription();
+            Debug.Log($"pc1 sdp: {desc.sdp}");
+
+            //var conf = pc1.GetConfiguration();
+            //Debug.Log($"ice servers count {conf.iceServers.Length}");
         }
     }
 
@@ -92,6 +74,36 @@ public class AddMediaStream : MonoBehaviour
         };
 
         return config;
+    }
+    IEnumerator Call()
+    {
+        callButton.interactable = false;
+        hangupButton.interactable = true;
+
+        Debug.Log("GetSelectedSdpSemantics");
+        var configuration = GetSelectedSdpSemantics();
+        pc1 = new RTCPeerConnection(ref configuration);
+        Debug.Log("Created local peer connection object pc1");
+        pc1.OnIceCandidate = pc1OnIceCandidate;
+        pc1.OnIceConnectionChange = pc1OnIceConnectionChange;
+        pc2 = new RTCPeerConnection(ref configuration);
+        Debug.Log("Created remote peer connection object pc2");
+        pc2.OnIceCandidate = pc2OnIceCandidate;
+        pc2.OnIceConnectionChange = pc2OnIceConnectionChange;
+
+        Debug.Log("pc1 createOffer start");
+        var op = pc1.CreateOffer(ref OfferOptions);
+        yield return op;
+
+        if (!op.isError)
+        {
+            yield return StartCoroutine(OnCreateOfferSuccess(op.desc));
+        }
+        else
+        {
+            OnCreateSessionDescriptionError(op.error);
+        }
+        sdpCheck = StartCoroutine(Loop());
     }
     void OnIceConnectionChange(RTCPeerConnection pc, RTCIceConnectionState state)
     {
@@ -125,118 +137,11 @@ public class AddMediaStream : MonoBehaviour
                 break;
         }
     }
-    IEnumerator Pc1OnNegotiationNeeded()
-    {
-        Debug.Log("pc1 createOffer start");
-        var op = pc1.CreateOffer(ref OfferOptions);
-        yield return op;
-
-        if (!op.isError)
-        {
-            yield return StartCoroutine(OnCreateOfferSuccess(op.desc));
-        }
-        else
-        {
-            OnCreateSessionDescriptionError(op.error);
-        }
-    }
-    void Pc1OnIceConnectinChange(RTCIceConnectionState state)
-    {
-        OnIceConnectionChange(pc1, state);
-    }
-    void Pc2OnIceConnectionChange(RTCIceConnectionState state)
-    {
-        OnIceConnectionChange(pc2, state);
-    }
-
-    void Pc1OnIceCandidate(RTCIceCandidate candidate)
-    {
-        OnIceCandidate(pc1, candidate);
-    }
-    void Pc2OnIceCandidate(RTCIceCandidate candidate)
-    {
-        OnIceCandidate(pc2, candidate);
-    }
-    public void AddTracks() 
-    {
-        foreach (var track in audioStream.GetTracks())
-        {
-            pc1Senders.Add (pc1.AddTrack(track));  
-        }
-        foreach(var track in videoStream.GetTracks())
-        {
-            pc1Senders.Add(pc1.AddTrack(track));
-        }
-        if(!videoUpdateStarted)
-        {
-            StartCoroutine(WebRTC.Update());
-            videoUpdateStarted = true;
-        }
-        addTracksButton.interactable = false;
-        removeTracksButton.interactable = true;
-    }
-
-    public void RemoveTracks()
-    {
-        foreach(var sender in pc1Senders)
-        {
-            pc1.RemoveTrack(sender);
-        }
-        foreach (var sender in pc2Senders)
-        {
-            pc2.RemoveTrack(sender);
-        }
-        pc1Senders.Clear();
-        pc2Senders.Clear();
-        addTracksButton.interactable = true;
-        removeTracksButton.interactable = false;
-        trackInfos.Clear();
-        infoText.text = "";
-    }
-
-    void Call()
-    {
-        callButton.interactable = false;
-        Debug.Log("GetSelectedSdpSemantics");
-        var configuration = GetSelectedSdpSemantics();
-        pc1 = new RTCPeerConnection(ref configuration);
-        Debug.Log("Created local peer connection object pc1");
-        pc1.OnIceCandidate = pc1OnIceCandidate;
-        pc1.OnIceConnectionChange = pc1OnIceConnectionChange;
-        pc1.OnNegotiationNeeded = pc1OnNegotiationNeeded;
-        pc2 = new RTCPeerConnection(ref configuration);
-        Debug.Log("Created remote peer connection object pc2");
-        pc2.OnIceCandidate = pc2OnIceCandidate;
-        pc2.OnIceConnectionChange = pc2OnIceConnectionChange;
-        pc2.OnTrack = pc2Ontrack;
-
-        RTCDataChannelInit conf = new RTCDataChannelInit(true);
-        dataChannel = pc1.CreateDataChannel("data", ref conf);
-        audioStream = Audio.CaptureStream();
-        videoStream = cam.CaptureStream(1280, 720);
-        RtImage.texture = cam.targetTexture;
- 
-    }
-
-    /// <summary>
-    /// 
-    /// </summary>
-    /// <param name="pc"></param>
-    /// <param name="streamEvent"></param>
     void OnIceCandidate(RTCPeerConnection pc, RTCIceCandidate​ candidate)
     {
         GetOtherPc(pc).AddIceCandidate(ref candidate);
         Debug.Log($"{GetName(pc)} ICE candidate:\n {candidate.candidate}");
     }
-    void OnTrack(RTCPeerConnection pc, RTCTrackEvent e)
-    {
-        pc2Senders.Add(pc.AddTrack(e.Track));
-        trackInfos.Append($"{GetName(pc)} receives remote track:\r\n");
-        trackInfos.Append($"Track kind: {e.Track.Kind}\r\n");
-        trackInfos.Append($"Track id: {e.Track.Id}\r\n");
-        infoText.text = trackInfos.ToString();
-    }
-
     string GetName(RTCPeerConnection pc)
     {
         return (pc == pc1) ? "pc1" : "pc2";
@@ -287,7 +192,7 @@ public class AddMediaStream : MonoBehaviour
         }
         else
         {
-            OnCreateSessionDescriptionError(op3.error); 
+            OnCreateSessionDescriptionError(op3.error);
         }
     }
 
@@ -333,18 +238,19 @@ public class AddMediaStream : MonoBehaviour
         }
     }
 
-    void OnAddIceCandidateSuccess(RTCPeerConnection pc)
-    {
-        Debug.Log($"{GetName(pc)} addIceCandidate success");
-    }
-
-    void OnAddIceCandidateError(RTCPeerConnection pc, RTCError error)
-    {
-        Debug.Log($"{GetName(pc)} failed to add ICE Candidate: ${error}");
-    }
-
     void OnCreateSessionDescriptionError(RTCError e)
     {
 
+    }
+    void Hangup()
+    {
+        Debug.Log("Ending call");
+        StopCoroutine(sdpCheck);
+        pc1.Close();
+        pc2.Close();
+        pc1 = null;
+        pc2 = null;
+        hangupButton.interactable = false;
+        callButton.interactable = true;
     }
 }
