+++ ProjectSettings/HDRPProjectSettings.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &1
+MonoBehaviour:
+  m_ObjectHideFlags: 61
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_GeneratorAsset: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 63a2978a97e4fc04cb9d905947216f3d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  version: 1
+  m_DefaultScenePrefabSaved: {fileID: 3628837357921210760, guid: 7a069bc3ea99ce14a9b310b38300d871,
+    type: 2}
+  m_ProjectSettingFolderPath: HDRPDefaultResources
+  m_PopupAtStart: 0
